+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Having types and fields
+- GroupType instead of anonymous declaration in GroupField
+
+### Changed
+- Depracated GroupField::FIELD_COLUMN constant, use GroupType::FIELD_COLUMN instead
 
 ## [0.2.1] - 2023-07-03
 ### Added
@@ -25,7 +31,8 @@
 - Webonyx driver
 - Webonyx schema transformer
 
-[Unreleased]: https://github.com/efabrica-team/graphql/compare/0.2.1...main
+[Unreleased]: https://github.com/efabrica-team/graphql/compare/0.2.2...main
+[0.2.2]: https://github.com/efabrica-team/graphql/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/efabrica-team/graphql/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/efabrica-team/graphql/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/efabrica-team/graphql/compare/0.0.0...0.1.0